rename update alert and add to license page
